+++ src/main/java/Main.java
@@ -2,7 +2,7 @@ import java.util.Scanner;
 
 public class Main {
     public static void main(String[] args) {
-        int liczba1, liczba2, wynikObu;
+        long liczba1, liczba2, wynikObu;
         Licz licz = new Licz();
         System.out.println("podaj liczbę");
         Scanner wczytywanie = new Scanner(System.in);
@@ -13,7 +13,7 @@ public class Main {
         System.out.println("podałeś " + liczba2);
         wynikObu = licz.sumuj(liczba1, liczba2);
         System.out.println("suma liczb " + liczba1 + " i liczby " + liczba2 + " to " + wynikObu);
-        int wynikKwadratu = licz.kwadratLiczby(liczba1);
+        long wynikKwadratu = licz.kwadratLiczby(liczba1);
         System.out.println(" kwadrat lioczby pierwszej podanej przez ciebie to " + wynikKwadratu);
         System.out.println("a kwadrat drugiej to " + licz.kwadratLiczby(liczba2));
         System.out.println("kwadrat wyniku wynosi " + licz.kwadratLiczby(wynikObu));
